fix: can not change edgeless mode after add blockhub setter
