+++ apps/web/src/pages/public-workspace/[workspaceId]/[pageId].tsx
@@ -89,9 +89,6 @@ const Page: NextPageWithLayout = () => {
               editor.readonly = true;
               setLoaded(true);
             }}
-            setBlockHub={() => {
-              return;
-            }}
           />
         )}
       </PageContainer>
