+++ apps/web/src/pages/workspace/[workspaceId]/[pageId].tsx
@@ -20,19 +20,43 @@ import { BlockHub } from '@blocksuite/blocks';
 const DynamicBlocksuite = dynamic(() => import('@/components/editor'), {
   ssr: false,
 });
+
+const BlockHubAppender = () => {
+  const { setBlockHub, editor } = useAppState();
+  const setBlockHubHandler = useCallback(
+    (blockHub: BlockHub) => setBlockHub.current(blockHub),
+    [setBlockHub]
+  );
+  useEffect(() => {
+    let blockHubElement: HTMLElement | null = null;
+
+    editor?.createBlockHub().then(blockHub => {
+      const toolWrapper = document.querySelector('#toolWrapper');
+      if (!toolWrapper) {
+        // In an invitation page there is no toolWrapper, which contains helper icon and blockHub icon
+        return;
+      }
+      blockHubElement = blockHub;
+      // setBlockHubHandler(blockHub);
+      toolWrapper.appendChild(blockHub);
+    });
+    return () => {
+      blockHubElement?.remove();
+    };
+  }, [editor, setBlockHubHandler]);
+  return null;
+};
+
 const Page: NextPageWithLayout = () => {
-  const { currentPage, currentWorkspace, setEditor, setBlockHub } =
-    useAppState();
+  const { currentPage, currentWorkspace, setEditor } = useAppState();
 
   const setEditorHandler = useCallback(
     (editor: EditorContainer) => setEditor.current(editor),
     [setEditor]
   );
-  const setBlockHubHandler = useCallback(
-    (blockHub: BlockHub) => setBlockHub.current(blockHub),
-    [setBlockHub]
-  );
+
   const { t } = useTranslation();
+
   return (
     <>
       <Head>
@@ -42,12 +66,14 @@ const Page: NextPageWithLayout = () => {
       <MobileModal />
 
       {currentPage && currentWorkspace?.blocksuiteWorkspace && (
-        <DynamicBlocksuite
-          page={currentPage}
-          workspace={currentWorkspace.blocksuiteWorkspace}
-          setEditor={setEditorHandler}
-          setBlockHub={setBlockHubHandler}
-        />
+        <>
+          <DynamicBlocksuite
+            page={currentPage}
+            workspace={currentWorkspace.blocksuiteWorkspace}
+            setEditor={setEditorHandler}
+          />
+          <BlockHubAppender />
+        </>
       )}
     </>
   );
